reorganize Monte2; remove Configuration dependence
