--- include/casm/monte2/io/json/Sampling_json_io.hh
@@ -1,8 +1,0 @@
-#ifndef CASM_monte2_Sampling_json_io
-#define CASM_monte2_Sampling_json_io
-
-namespace CASM {
-namespace Monte2 {}  // namespace Monte2
-}  // namespace CASM
-
-#endif
